+++ rapdev_validator/datadog_checks/rapdev_validator/config_models/defaults.py
@@ -4,7 +4,23 @@
 from datadog_checks.base.utils.models.fields import get_default_field_value
 
 
+def shared_api_key(field, value):
+    return False
+
+
+def shared_app_key(field, value):
+    return get_default_field_value(field, value)
+
+
 def shared_service(field, value):
+    return get_default_field_value(field, value)
+
+
+def instance_api_key(field, value):
+    return False
+
+
+def instance_app_key(field, value):
     return get_default_field_value(field, value)